# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Planet Path API
  description: RESTful API for Planet Path - An environmental education platform
  version: 2.0.0
  contact:
    name: Planet Path API Support
    email: support@planetpath.com
servers:
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/courses:
    get:
      tags:
        - Courses
      summary: Get all courses
      description: Retrieve a paginated list of courses with optional filtering
      operationId: getCourses
      parameters:
        - name: status
          in: query
          description: Filter by course status
          required: false
          schema:
            type: string
            enum: [draft, published, archived]
            default: published
        - name: impact_type
          in: query
          description: Filter by impact type
          required: false
          schema:
            type: string
            enum: [climate, waste, energy, water, community, other]
        - name: tags
          in: query
          description: Filter by tags (can be multiple)
          required: false
          schema:
            type: array
            items:
              type: string
        - name: authorId
          in: query
          description: Filter by author ID
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of courses per page
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort
          in: query
          description: Sort field (prefix with - for descending)
          required: false
          schema:
            type: string
            default: -createdAt
      responses:
        '200':
          description: List of courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoursesListResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Courses
      summary: Create a new course
      description: Create a new course (instructor/admin only)
      operationId: createCourse
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateInput'
            example:
              title: Climate Change Basics
              description: Learn the fundamentals of climate change and its impact on our planet
              modules: []
              tags:
                - climate
                - environment
                - basics
              price: 0
              impact_type: climate
              status: draft
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseResponse'
        '400':
          description: Validation error or slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - User does not have required role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/lessons:
    post:
      tags:
        - Lessons
      summary: Create a new lesson
      description: Create a new lesson within a module
      operationId: createLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonCreateInput'
            example:
              title: Introduction to Climate Science
              content: This lesson covers the basic principles of climate science...
              videoUrl: https://example.com/video.mp4
              duration: 15
              order: 1
              moduleId: 507f1f77bcf86cd799439011
      responses:
        '201':
          description: Lesson created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Module not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Lessons
      summary: Get lessons by module
      description: Retrieve all lessons for a specific module
      operationId: getLessonsByModule
      parameters:
        - name: moduleId
          in: query
          required: true
          description: Module ID
          schema:
            type: string
      responses:
        '200':
          description: Lessons retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      lessons:
                        type: array
                        items:
                          $ref: '#/components/schemas/LessonResponse'
                      count:
                        type: integer
                        example: 5
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/assignments:
    post:
      tags:
        - Assignments
      summary: Create a new assignment
      description: Create a new assignment (instructor/admin only)
      operationId: createAssignment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCreateInput'
            example:
              title: Climate Action Project
              description: Create a project demonstrating climate action in your community
              courseId: 507f1f77bcf86cd799439011
              dueDate: '2024-12-31T23:59:59.000Z'
              maxScore: 100
              attachments: []
      responses:
        '201':
          description: Assignment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Instructor or admin role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course or lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Assignments
      summary: Get assignment by ID
      description: Retrieve a single assignment by ID
      operationId: getAssignment
      parameters:
        - name: id
          in: path
          required: true
          description: Assignment ID
          schema:
            type: string
      responses:
        '200':
          description: Assignment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentResponse'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/submissions:
    post:
      tags:
        - Submissions
      summary: Submit an assignment
      description: Submit an assignment with files and metadata (multipart/form-data)
      operationId: submitAssignment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - assignmentId
                - courseId
              properties:
                assignmentId:
                  type: string
                  description: Assignment ID
                courseId:
                  type: string
                  description: Course ID
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Array of files to upload
                metadata:
                  type: string
                  description: JSON string containing metadata (geotag, notes)
                  example: '{"geotag":{"lat":40.7128,"lng":-74.0060},"notes":"Project completed at Central Park"}'
            encoding:
              files:
                contentType: application/pdf, image/*, video/*
              metadata:
                contentType: application/json
      responses:
        '201':
          description: Assignment submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '400':
          description: Validation error or already submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: User's full name
          example: John Doe
        email:
          type: string
          format: email
          description: User's email address
          example: john.doe@example.com
        role:
          type: string
          enum: [student, instructor, admin]
          description: User's role
          example: student
        xp:
          type: integer
          minimum: 0
          description: User's experience points
          example: 100
        badges:
          type: array
          items:
            type: string
          description: User's badges
          example:
            - eco-warrior
            - climate-champion
    CourseCreateInput:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: Course title
          example: Climate Change Basics
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          description: URL-friendly course identifier (auto-generated if not provided)
          example: climate-change-basics
        description:
          type: string
          minLength: 10
          description: Course description
          example: Learn the fundamentals of climate change and its impact on our planet
        modules:
          type: array
          items:
            type: string
          description: Array of module ObjectIds
          example: []
        tags:
          type: array
          maxItems: 10
          items:
            type: string
          description: Course tags
          example:
            - climate
            - environment
            - basics
        price:
          type: number
          minimum: 0
          default: 0
          description: Course price
          example: 0
        impact_type:
          type: string
          enum: [climate, waste, energy, water, community, other]
          default: climate
          description: Type of environmental impact
          example: climate
        status:
          type: string
          enum: [draft, published, archived]
          default: draft
          description: Course status
          example: draft
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
          description: Course ID
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          description: Course title
          example: Climate Change Basics
        slug:
          type: string
          description: URL-friendly course identifier
          example: climate-change-basics
        description:
          type: string
          description: Course description
          example: Learn the fundamentals of climate change
        authorId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/UserResponse'
          description: Course author (populated with user details when fetched)
        modules:
          type: array
          items:
            type: string
          description: Array of module ObjectIds
        tags:
          type: array
          items:
            type: string
          description: Course tags
          example:
            - climate
            - environment
        price:
          type: number
          minimum: 0
          description: Course price
          example: 0
        impact_type:
          type: string
          enum: [climate, waste, energy, water, community, other]
          description: Type of environmental impact
          example: climate
        status:
          type: string
          enum: [draft, published, archived]
          description: Course status
          example: published
        createdAt:
          type: string
          format: date-time
          description: Course creation timestamp
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          description: Course last update timestamp
          example: '2024-01-01T00:00:00.000Z'
    CoursesListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            courses:
              type: array
              items:
                $ref: '#/components/schemas/CourseResponse'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  description: Current page number
                  example: 1
                limit:
                  type: integer
                  description: Number of items per page
                  example: 10
                total:
                  type: integer
                  description: Total number of courses
                  example: 1
                pages:
                  type: integer
                  description: Total number of pages
                  example: 1
...
    LessonCreateInput:
      type: object
      required:
        - title
        - content
        - order
        - moduleId
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: Lesson title
          example: Introduction to Climate Science
        content:
          type: string
          description: Lesson content/text
          example: This lesson covers the basic principles of climate science...
        videoUrl:
          type: string
          format: uri
          description: URL to lesson video
          example: https://example.com/video.mp4
        duration:
          type: integer
          minimum: 0
          description: Lesson duration in minutes
          example: 15
        order:
          type: integer
          minimum: 1
          description: Lesson order in module
          example: 1
        moduleId:
          type: string
          description: Module ID
          example: 507f1f77bcf86cd799439011
    LessonResponse:
      type: object
      properties:
        id:
          type: string
          description: Lesson ID
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          description: Lesson title
          example: Introduction to Climate Science
        content:
          type: string
          description: Lesson content
        videoUrl:
          type: string
          format: uri
          description: Video URL
        duration:
          type: integer
          description: Duration in minutes
        order:
          type: integer
          description: Lesson order
        moduleId:
          type: string
          description: Module ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AssignmentCreateInput:
      type: object
      required:
        - title
        - description
        - dueDate
        - maxScore
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          description: Assignment title
          example: Climate Action Project
        description:
          type: string
          description: Assignment description
          example: Create a project demonstrating climate action in your community
        lessonId:
          type: string
          description: Lesson ID (optional, either lessonId or courseId required)
          example: 507f1f77bcf86cd799439011
        courseId:
          type: string
          description: Course ID (optional, either lessonId or courseId required)
          example: 507f191e810c19729de860ea
        dueDate:
          type: string
          format: date-time
          description: Assignment due date (must be in future)
          example: '2024-12-31T23:59:59.000Z'
        maxScore:
          type: integer
          minimum: 1
          default: 100
          description: Maximum score for assignment
          example: 100
        attachments:
          type: array
          items:
            type: string
          description: Array of attachment URLs
          example: []
    AssignmentResponse:
      type: object
      properties:
        id:
          type: string
          description: Assignment ID
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          description: Assignment title
        description:
          type: string
          description: Assignment description
        lessonId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/LessonResponse'
          description: Lesson ID or populated lesson
        courseId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/CourseResponse'
          description: Course ID or populated course
        dueDate:
          type: string
          format: date-time
          description: Due date
        maxScore:
          type: integer
          description: Maximum score
          example: 100
        attachments:
          type: array
          items:
            type: string
          description: Attachment URLs
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubmissionMetadata:
      type: object
      properties:
        geotag:
          type: object
          properties:
            lat:
              type: number
              minimum: -90
              maximum: 90
              description: Latitude
              example: 40.7128
            lng:
              type: number
              minimum: -180
              maximum: 180
              description: Longitude
              example: -74.0060
        notes:
          type: string
          maxLength: 1000
          description: Submission notes
          example: Project completed at Central Park
...
    SubmissionResponse:
      type: object
      properties:
        id:
          type: string
          description: Submission ID
          example: 507f1f77bcf86cd799439011
        assignmentId:
          type: string
          description: Assignment ID
        courseId:
          type: string
          description: Course ID
        userId:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/UserResponse'
          description: User ID or populated user
        files:
          type: array
          items:
            type: string
          description: Array of file URLs
          example:
            - https://storage.example.com/submissions/file1.pdf
        metadata:
          $ref: '#/components/schemas/SubmissionMetadata'
        score:
          type: number
          nullable: true
          description: Assignment score
          example: 85
        feedback:
          type: string
          description: Instructor feedback
          example: Great work! Well documented project.
        status:
          type: string
          enum: [submitted, graded, rejected]
          description: Submission status
          example: submitted
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name with error
                example: email
              message:
                type: string
                description: Error message
                example: Please provide a valid email address
          description: Array of validation errors (only present for validation errors)
        error:
          type: string
          description: Detailed error message (only in development mode)
          example: Detailed error message here